fix(bookmark): prevent twitter emoji from being displayed as image
Some bookmarks, like those for twitter profiles, were showing an emoji as the bookmark image. This was not the desired behavior.

This commit adds a check to prevent images from `twimg.com/emoji` from being displayed as bookmark images. The comment has been updated to be more specific about this case.